src/third_party/blink/web_tests/fast/html/imports/sub-imports-onload-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and will be removed in M73, around March 2019. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 for more details.
This ensures that load and error events are fired on sub-imports.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS theEvent.type is 'error'
PASS theEvent.target.ownerDocument is root.import
PASS theEvent.type is 'load'
PASS theEvent.target.ownerDocument is root.import
PASS successfullyParsed is true

TEST COMPLETE
